feat: add combined 0xB0CD and 0xB826 hexdump type
This new type utilises the existing 0xB0CD and 0xB826 hexdump parsers
by combining their output into a single output and parser run.

This simplifies the usage of the parser when using Qualcomm QCAT
outputs, and prevents the need to run the parser twice to get a full
set of capabilities.

The new type should be used in combination with `-input` (0xB0CD) and
`-inputNR` (0xB826) options, as well as `--multiple0xB826` (if desired).