utility: bwstring: output both UL and DL bws
The current implementation reports UL BWs for SDL combos and DL BWs for other combos.

First, the reporting of UL BWs for SDL bands was incorrect, but this was due to a bug in the ComponentNr#isSUL function.
Then some bands have different BWs between DL and UL.

So let's report both and fix both of issues.