importcapabilityinformationjson: set "default" 100MHz even if no bw is reported for 30kHz
According to TS 38.306 v16.6.0 there's no 100MHz field for n41, n48, n77, n78, n79, n90.
So we assumed that it was supported by default, except if no bw was reported for that scs.
Remove that check for 30kHz scs, since there are devices that report no supported bw for those bands.