feat: add keyed json output option
This output format produces a JSON-compliant output containing an object that maps the capability type to its CSV data.

This makes it easier to handle the outputs from multiple parses (e.g., 0xB0CD plus 0xB826) and merge them together to allow for a simpler import process for a linked application.

This also prevents confusion between the various NR outputs (EN-DC, NR, NR-DC) as some devices may, for example, support NR SA but not ENDC, and currently there is no way for an automated process to determine which types of CSV have been outputted.